Move query api tests to its own directory
